Supporting postgres

+++ src/main/java/BusinessLayer/Policies/DiscountPolicies/PolicyTypes/MinBagTotalDiscountPolicy.java
@@ -17,7 +17,7 @@ public class MinBagTotalDiscountPolicy extends BaseDiscountPolicy {
     private final double minTotal;
 
     public MinBagTotalDiscountPolicy(int policyId, double minTotal,int store_id,int discount_id) throws Exception {
-        super(policyId,store_id,discount_id);
+        super(policyId,store_id, (long) discount_id);
         if (minTotal <= 0) {
             logger.error("minimum total must be larger than 0 but is " + minTotal);
             throw new Exception("minimum total must be larger than 0");
