+++ src/main/java/Repositories/IDiscountRepo.java
@@ -1,7 +1,6 @@
 package Repositories;
 
 import BusinessLayer.Discounts.Discount;
-import Notification.Notification;
 
 import java.util.concurrent.ConcurrentLinkedQueue;
 
@@ -9,7 +8,7 @@ public interface IDiscountRepo {
     void addDiscount(Discount notification);
     void removeDiscount(Discount notification);
 
-    Discount get(Integer id);
+    Discount get(Long id);
     void clear();
     ConcurrentLinkedQueue<Discount> getAllDiscounts();
 }
